LPS-54457 Move replicate classes used by hibernate from com.liferay.portal.cache.cluster to com.liferay.portal.dao.orm.hibernate.region
